Refactor values
